refactor: Moving formz inputs to validators rather than models and remove relative import
