+++ src/support_sphere/lib/logic/cubit/login_cubit.dart
@@ -1,7 +1,8 @@
 import 'package:support_sphere/data/repositories/authentication.dart';
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
-import 'package:support_sphere/data/models/all_models.dart';
+import 'package:support_sphere/data/validators/login/email.dart';
+import 'package:support_sphere/data/validators/login/password.dart';
 import 'package:support_sphere/logic/cubit/utils.dart';
 import 'package:formz/formz.dart';
